requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Jinja2==2.11.2
python-dotenv===0.14.0
Werkzeug===1.0.1
pymysql
sqlalchemy